ai_services4/audio-emotion/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
transformers==4.35.2
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1
python-dotenv==1.0.0
pydantic==2.5.0